React: move all the hooks into hooks.ts
