+++ packages/connect-react/src/create-app-hook.ts
@@ -1,98 +1,7 @@
-import { useEffect, useMemo, useRef, useState } from 'react'
-import { SubscriptionHandler, SubscriptionResult } from '@aragon/connect-types'
+import { useEffect, useRef, useState } from 'react'
+import { SubscriptionResult } from '@aragon/connect-types'
 import { App } from '@aragon/connect'
-import { LoadingStatus } from './types'
-
-type AppDataState<T> = {
-  error: Error | null
-  loading: boolean
-  result?: T
-}
-
-const appDataDefaultState = {
-  error: null,
-  loading: true,
-  result: undefined,
-}
-
-export function useConnectData<T = any>(
-  callback?: () => T | Promise<T> | SubscriptionResult<T>,
-  dependencies?: any[]
-): [T | undefined, LoadingStatus] {
-  const [{ result, error, loading }, setStatus] = useState<AppDataState<T>>(
-    appDataDefaultState
-  )
-
-  const callbackRef = useRef<Function>(() => undefined)
-  useEffect(() => {
-    callbackRef.current = callback || (() => undefined)
-  }, [callback])
-
-  const previousResultJsonRef = useRef<string>('')
-
-  useEffect(() => {
-    let cancelled = false
-    let subscriptionHandler: SubscriptionHandler | null = null
-
-    const update = async () => {
-      // keep the result until the next update
-      setStatus((status) => ({ ...status, error: null, loading: true }))
-
-      try {
-        const result = await callbackRef.current()
-
-        if (cancelled || result === undefined) {
-          return
-        }
-
-        // Subscription
-        if (typeof result === 'function') {
-          subscriptionHandler = result((error: Error | null, result?: any) => {
-            if (cancelled) {
-              return
-            }
-
-            if (error) {
-              setStatus({ error, loading: false, result: undefined })
-              previousResultJsonRef.current = JSON.stringify(null)
-              return
-            }
-
-            const resultJson = JSON.stringify(result)
-            if (resultJson === previousResultJsonRef.current) {
-              // no update
-              return
-            }
-
-            previousResultJsonRef.current = resultJson
-            setStatus({ error: null, loading: false, result })
-          })
-
-          // Just async data
-        } else {
-          setStatus({ error: null, loading: false, result })
-        }
-      } catch (err) {
-        if (!cancelled) {
-          setStatus({ error: err, loading: false, result: undefined })
-        }
-      }
-    }
-    update()
-
-    return () => {
-      cancelled = true
-      subscriptionHandler?.unsubscribe()
-    }
-  }, dependencies || [])
-
-  const status = useMemo(() => ({ error, loading, retry: () => null }), [
-    error,
-    loading,
-  ])
-
-  return [result, status]
-}
+import { useConnectData } from './hooks'
 
 export function createAppHook(
   appConnect: Function,
