+++ packages/connect-react/src/hooks.ts
@@ -1,8 +1,103 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
-import { AppFiltersParam, SubscriptionHandler } from '@aragon/connect-types'
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import {
+  AppFiltersParam,
+  SubscriptionHandler,
+  SubscriptionResult,
+} from '@aragon/connect-types'
 import { App, Organization, Permission } from '@aragon/connect'
 import { LoadingStatus, OrganizationHookResult } from './types'
 import { useConnectContext } from './connect-context'
+
+type AppDataState<T> = {
+  error: Error | null
+  loading: boolean
+  result?: T
+}
+
+const appDataDefaultState = {
+  error: null,
+  loading: true,
+  result: undefined,
+}
+
+export function useConnectData<T = any>(
+  callback?: () => T | Promise<T> | SubscriptionResult<T>,
+  dependencies?: any[]
+): [T | undefined, LoadingStatus] {
+  const [{ result, error, loading }, setStatus] = useState<AppDataState<T>>(
+    appDataDefaultState
+  )
+
+  const callbackRef = useRef<Function>(() => undefined)
+  useEffect(() => {
+    callbackRef.current = callback || (() => undefined)
+  }, [callback])
+
+  const previousResultJsonRef = useRef<string>('')
+
+  useEffect(() => {
+    let cancelled = false
+    let subscriptionHandler: SubscriptionHandler | null = null
+
+    const update = async () => {
+      // keep the result until the next update
+      setStatus((status) => ({ ...status, error: null, loading: true }))
+
+      try {
+        const result = await callbackRef.current()
+
+        if (cancelled || result === undefined) {
+          return
+        }
+
+        // Subscription
+        if (typeof result === 'function') {
+          subscriptionHandler = result((error: Error | null, result?: any) => {
+            if (cancelled) {
+              return
+            }
+
+            if (error) {
+              setStatus({ error, loading: false, result: undefined })
+              previousResultJsonRef.current = JSON.stringify(null)
+              return
+            }
+
+            const resultJson = JSON.stringify(result)
+            if (resultJson === previousResultJsonRef.current) {
+              // no update
+              return
+            }
+
+            previousResultJsonRef.current = resultJson
+            setStatus({ error: null, loading: false, result })
+          })
+
+          // Just async data
+        } else {
+          setStatus({ error: null, loading: false, result })
+        }
+      } catch (err) {
+        if (!cancelled) {
+          setStatus({ error: err, loading: false, result: undefined })
+        }
+      }
+    }
+    update()
+
+    return () => {
+      cancelled = true
+      subscriptionHandler?.unsubscribe()
+    }
+  }, dependencies || [])
+
+  const status = useMemo(() => ({ error, loading, retry: () => null }), [
+    error,
+    loading,
+  ])
+
+  return [result, status]
+}
 
 export function useSubscription<Data>(
   callback: (
@@ -108,13 +203,17 @@ export function useApp(
 
 export function useApps(appsFilter?: AppFiltersParam): [App[], LoadingStatus] {
   const callback = useCallback(
-    (org, onData) => org.onApps(appsFilter, onData),
+    (org, onData) => {
+      return org.onApps(appsFilter, onData)
+    },
     [JSON.stringify(appsFilter)]
   )
   return useSubscription<App[]>(callback, [])
 }
 
 export function usePermissions(): [Permission[], LoadingStatus] {
-  const callback = useCallback((org, onData) => org.onPermissions(onData), [])
+  const callback = useCallback((org, onData) => {
+    return org.onPermissions(onData)
+  }, [])
   return useSubscription<Permission[]>(callback, [])
 }
