+++ packages/connect-react/src/index.ts
@@ -1,4 +1,4 @@
 export * from '@aragon/connect'
 export * from './connect-provider'
 export * from './create-app-hook'
-export * from './org-hooks'
+export * from './hooks'
